+++ modules/contracts/src.ts/services/ethService.spec.ts
@@ -39,6 +39,7 @@ let getCodeMock: SinonStub;
 let getOnchainBalanceMock: SinonStub;
 let waitForConfirmation: SinonStub<[chainId: number, responses: TransactionResponse[]], Promise<TransactionReceipt>>;
 let getGasPrice: SinonStub<[chainId: number], Promise<Result<BigNumber, ChainError>>>;
+let provider1: SinonStubbedInstance<JsonRpcProvider>;
 
 let channelState: FullChannelState;
 
@@ -98,6 +99,7 @@ describe("ethService unit test", () => {
 
     const _provider = createStubInstance(JsonRpcProvider);
     _provider.getTransaction.resolves(txResponse);
+    provider1 = _provider;
     provider1337 = _provider;
     provider1338 = _provider;
     (signer as any).provider = provider1337;
@@ -108,6 +110,7 @@ describe("ethService unit test", () => {
       {
         1337: provider1337,
         1338: provider1338,
+        1: provider1,
       },
       signer,
       log,
@@ -638,12 +641,12 @@ describe("ethService unit test", () => {
     });
 
     it("should wait for the required amount of confirmations", async () => {
-      provider1337.getTransactionReceipt.onFirstCall().resolves({ ...txReceipt, confirmations: 0 });
-      provider1337.getTransactionReceipt.onSecondCall().resolves({ ...txReceipt, confirmations: 0 });
-      provider1337.getTransactionReceipt.onThirdCall().resolves(txReceipt);
-      const res = await ethService.waitForConfirmation(1337, [txResponse]);
+      provider1.getTransactionReceipt.onFirstCall().resolves({ ...txReceipt, confirmations: 0 });
+      provider1.getTransactionReceipt.onSecondCall().resolves({ ...txReceipt, confirmations: 0 });
+      provider1.getTransactionReceipt.onThirdCall().resolves(txReceipt);
+      const res = await ethService.waitForConfirmation(1, [txResponse]);
       expect(res).to.deep.eq(txReceipt);
-      expect(provider1337.getTransactionReceipt.callCount).to.eq(3);
+      expect(provider1.getTransactionReceipt.callCount).to.eq(3);
     });
 
     it("should error with a timeout error if it is past the confirmation time", async () => {