+++ modules/contracts/src.ts/services/ethService.ts
@@ -23,8 +23,7 @@ import {
   encodeTransferResolver,
   encodeTransferState,
   getRandomBytes32,
-  generateMerkleTreeData,
-  hashCoreTransferState,
+  getMerkleProof,
 } from "@connext/vector-utils";
 import { Signer } from "@ethersproject/abstract-signer";
 import { BigNumber } from "@ethersproject/bignumber";
@@ -74,8 +73,9 @@ export const waitForTransaction = async (
 };
 
 export class EthereumChainService extends EthereumChainReader implements IVectorChainService {
+  private nonces: Map<number, number> = new Map();
   private signers: Map<number, Signer> = new Map();
-  private queue: PriorityQueue = new PriorityQueue({ concurrency: 1 });
+  private queues: Map<number, PriorityQueue> = new Map();
   private evts: { [eventName in ChainServiceEvent]: Evt<ChainServiceEventMap[eventName]> } = {
     ...this.disputeEvts,
     [ChainServiceEvents.TRANSACTION_SUBMITTED]: new Evt(),
@@ -95,6 +95,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
         parseInt(chainId),
         typeof signer === "string" ? new Wallet(signer, provider) : (signer.connect(provider) as Signer),
       );
+      this.queues.set(parseInt(chainId), new PriorityQueue({ concurrency: 1 }));
     });
 
     // TODO: Check to see which tx's are still active / unresolved, and resolve them.
@@ -195,21 +196,36 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     txFn: (gasPrice: BigNumber, nonce: number) => Promise<TransactionResponse | undefined>,
     gasPrice: BigNumber,
     signer: Signer,
+    chainId: number,
     nonce?: number,
   ): Promise<Result<TransactionResponse | undefined, Error>> {
     // Queue up the execution of the transaction.
-    return await this.queue.add(
-      async (): Promise<Result<TransactionResponse | undefined, Error>> => {
-        try {
-          // Send transaction using the passed in callback.
-          const actualNonce: number = nonce ?? (await signer.getTransactionCount("pending"));
-          const response: TransactionResponse | undefined = await txFn(gasPrice, actualNonce);
-          return Result.ok(response);
-        } catch (e) {
-          return Result.fail(e);
+    if (!this.queues.has(chainId)) {
+      return Result.fail(new ChainError(ChainError.reasons.SignerNotFound));
+    }
+    // Define task to send tx with proper nonce
+    const task = async (): Promise<Result<TransactionResponse | undefined, Error>> => {
+      try {
+        // Send transaction using the passed in callback.
+        const stored = this.nonces.get(chainId);
+        const nonceToUse: number = nonce ?? stored ?? (await signer.getTransactionCount("pending"));
+        const response: TransactionResponse | undefined = await txFn(gasPrice, nonceToUse);
+        // After calling tx fn, set nonce to the greatest of
+        // stored, pending, or incremented
+        const pending = await signer.getTransactionCount("pending");
+        const incremented = (response?.nonce ?? nonceToUse) + 1;
+        // Ensure the nonce you store is *always* the greatest of the values
+        const toCompare = stored ?? 0;
+        if (toCompare < pending || toCompare < incremented) {
+          this.nonces.set(chainId, incremented > pending ? incremented : pending);
         }
-      },
-    );
+        return Result.ok(response);
+      } catch (e) {
+        return Result.fail(e);
+      }
+    };
+    const result = await this.queues.get(chainId)!.add(task);
+    return result;
   }
 
   /// Check to see if any txs were left in an unfinished state. This should only execute on
@@ -358,6 +374,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     // in this data with the already-existing store record of the tx.
     let responses: TransactionResponse[] = [];
     let nonce: number | undefined;
+    let nonceExpired: boolean = false;
     let receipt: TransactionReceipt | undefined;
     let gasPrice: BigNumber;
 
@@ -389,7 +406,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
           { method, methodId, nonce, tryNumber, channelAddress, gasPrice: gasPrice.toString() },
           "Attempting to send transaction",
         );
-        const result = await this.sendTx(txFn, gasPrice, signer, nonce);
+        const result = await this.sendTx(txFn, gasPrice, signer, chainId, nonce);
         if (!result.isError) {
           const response = result.getValue();
           if (response) {
@@ -456,6 +473,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
               // Another ethers message that we could potentially be getting back.
               error.message.includes("There is another transaction with same nonce in the queue."))
           ) {
+            nonceExpired = true;
             this.log.info(
               { method, methodId, channelAddress, reason, nonce, error: error.message },
               "Nonce already used: proceeding to check for confirmation in previous transactions.",
@@ -479,6 +497,23 @@ export class EthereumChainService extends EthereumChainReader implements IVector
       } catch (e) {
         // Check if the error was a confirmation timeout.
         if (e.message === ChainError.reasons.ConfirmationTimeout) {
+          if (nonceExpired) {
+            const error = new ChainError(ChainError.reasons.NonceExpired, {
+              methodId,
+              method,
+            });
+            await this.handleTxFail(
+              onchainTransactionId,
+              method,
+              methodId,
+              channelAddress,
+              reason,
+              receipt,
+              error,
+              "Nonce expired and could not confirm tx",
+            );
+            return Result.fail(error);
+          }
           // Scale up gas by percentage as specified by GAS_BUMP_PERCENT.
           // From ethers docs:
           // Generally, the new gas price should be about 50% + 1 wei more, so if a gas price
@@ -1329,7 +1364,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     }
 
     // Generate merkle root
-    const { tree } = generateMerkleTreeData(activeTransfers);
+    const proof = getMerkleProof(activeTransfers, transferIdToDispute);
 
     const res = await this.sendTxWithRetries(
       transferState.channelAddress,
@@ -1337,7 +1372,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
       TransactionReason.disputeTransfer,
       (gasPrice, nonce) => {
         const channel = new Contract(transferState.channelAddress, VectorChannel.abi, signer);
-        return channel.disputeTransfer(transferState, tree.getHexProof(hashCoreTransferState(transferState)), {
+        return channel.disputeTransfer(transferState, proof, {
           gasPrice,
           nonce,
         });